Fixes for linux

+++ VroomEditor/src/TestScene/RotatingDirLightScript.cpp
@@ -1,5 +1,7 @@
 #include "VroomEditor/TestScene/RotatingDirLightScript.h"
 
+#include <cmath>
+
 #include "Vroom/Scene/Entity.h"
 #include "Vroom/Scene/Components/TransformComponent.h"
 
@@ -13,7 +15,7 @@ void RotatingDirLightScript::onSpawn()
 void RotatingDirLightScript::onUpdate(float dt)
 {
   m_angle += dt * m_rotSpeed;
-  m_angle = std::fmodf(m_angle, glm::radians(360.f));
+  m_angle = std::fmod(m_angle, glm::radians(360.f));
   auto& tc = getEntity().getComponent<vrm::TransformComponent>();
   auto rot = tc.getRotation();
   rot.z = m_angle;
